- Grouped game types - Added game folder to hold game types
