--- src/main/model/Plinko/Direction.ts
@@ -1,4 +1,0 @@
-export enum Direction {
-  LEFT = 'LEFT',
-  RIGHT = 'RIGHT'
-}
