--- src/main/model/Plinko/Drop.ts
@@ -1,7 +1,0 @@
-import { Direction } from "./Direction";
-
-export interface Drop {
-  directions: Direction[],
-  multiplierRow: number[],
-  index: number
-}
