--- src/main/model/Plinko/Rows.ts
@@ -1,11 +1,0 @@
-export enum Rows {
-  EIGHT = 8,
-  NINE = 9,
-  TEN = 10,
-  ELEVEN = 11,
-  TWELVE = 12,
-  THIRTEEN = 13,
-  FOURTEEN = 14,
-  FIFTEEN = 15,
-  SIXTEEN = 16
-}
