--- src/main/model/Wheel/Spin.ts
@@ -1,4 +1,0 @@
-export interface Spin {
-  multipliers: number[],
-  index: number
-}
