+++ src/main/model/game/DiamondsTypes.ts
