+++ src/main/model/game/ScarabSpinTypes.ts
