+++ src/test/Wheel.test.ts
@@ -3,8 +3,7 @@ import verifyWheel from "../main/Wheel"
 import { SingleResultGameEvent } from "../main/model/GameEvent"
 import { GameSeed } from "../main/model/GameSeed"
 import { Risk } from "../main/model/Risk"
-import { Segments } from "../main/model/Wheel/Segments"
-import { Spin } from "../main/model/Wheel/Spin"
+import { Segments, Spin } from "../main/model/game/WheelTypes"
 
 import GameSeeds from "./resources/GameSeeds.json"
 import WheelMultiplierTable from "../main/resources/WheelMultiplierTable.json"
